Added negatives test, fixed testing output

+++ SortingClassLibrary/src/SortLib.cpp
@@ -4,11 +4,11 @@
 #include <array>
 
 
-void SelectionSort(std::vector<int>& a_in) {
+void SelectionSort(std::vector<double>& a_in) {
 	for (unsigned int outer = 0; outer < (a_in.size() - 1); outer++) {
 		for (unsigned int inner = outer + 1; inner < a_in.size(); inner++) {
 			if (a_in[inner] < a_in[outer]) {
-				int temp = a_in[inner];
+				double temp = a_in[inner];
 				a_in[inner] = a_in[outer];
 				a_in[outer] = temp;
 			}
